LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/pointerevents/pointerevent_capture_suppressing_mouse-expected.txt



PASS Validate pointer events track pointer movement without pointer capture.
FAIL Test pointer capture. assert_equals: Check after button hover: Unexpected events pointermove@captureButton expected 0 but got 1
